pyproject.toml
[tool.poetry]
name = "FADR"
version = "1.0.0"
description = "Fully Automated Disc Ripper (FADR)"
authors = ["jochenp <jochenp@gmail.com>"]
license = "MIT License"
packages = [
    { include = "fadr", from = "." },
]
readme = "README.md"
keywords = ["fadr", "python3"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
python = "^3.11"
alembic = "==1.14.1"
apprise = "==1.7.6"
argparse = "==1.4.0"
bcrypt = "==3.2.2" # Newer versions dont have Wheel
beautifulsoup4 = "==4.12.3"
certifi = "==2024.12.14"
cffi = "==1.16.0"
charset-normalizer = "==3.4.1"
click = "==8.1.7"
colorama = "==0.4.6"
discid = "==1.2.0"
flask = "==2.2.5"
flask-cors = "==4.0.0"
flask-login = "==0.6.3"
flask-migrate = "==4.0.4"
flask-sqlalchemy = "==3.1.1"
flask-wtf = "==1.1.1"
greenlet = "==3.1.1"
idna = "==3.10"
itsdangerous = "==2.1.2"
jinja2 = "==3.1.2"
mako = "==1.2.4"
markdown = "==3.7"
markupsafe = "==2.1.3"
musicbrainzngs = "==0.7.1"
netifaces = "==0.11.0"
oauthlib = "==3.2.2"
prettytable = "==3.10.0"
psutil = "==6.1.1"
pycparser = "==2.21"
pycurl = "==7.45.6"
pydvdid = "==1.1"
pyudev = "==0.24.3"
pyyaml = "==6.0.2"
requests = "==2.32.3"
requests-oauthlib = "==1.3.1"
robobrowser = "==0.5.3"
six = "==1.16.0"
soupsieve = "==2.5"
sqlalchemy = "==2.0.19"
tinydownload = "==0.1.0"
urllib3 = "==2.3.0"
waitress = "==2.1.2"
wcwidth = "==0.2.13"
werkzeug = "==2.3.6"
wtforms = "==3.0.1"
xmltodict = "==0.13.0"


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.1.2"
isort = "^6.0.1"
pre-commit = "^4.1.0"
pyupgrade = "^3.19.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
